7709 - Register and reference everything related to DynamoDb
