+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Application.CommandHandlers/xxENSONOxx.xxSTACKSxx.Application.CommandHandlers.csproj
@@ -8,6 +8,6 @@
     <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.CQRS\xxENSONOxx.xxSTACKSxx.CQRS.csproj" />
     <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Domain\xxENSONOxx.xxSTACKSxx.Domain.csproj" />
     <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus.csproj" />
-    <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.csproj" />
+    <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Infrastructure\xxENSONOxx.xxSTACKSxx.Infrastructure.csproj" />
   </ItemGroup>
 </Project>
