+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Application.QueryHandlers/xxENSONOxx.xxSTACKSxx.Application.QueryHandlers.csproj
@@ -8,7 +8,7 @@
         <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Application.Integration\xxENSONOxx.xxSTACKSxx.Application.Integration.csproj" />
         <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.CQRS\xxENSONOxx.xxSTACKSxx.CQRS.csproj" />
         <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions.csproj" />
-        <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.csproj" />
+        <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Infrastructure\xxENSONOxx.xxSTACKSxx.Infrastructure.csproj" />
     </ItemGroup>
 
 </Project>
