+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Domain/xxENSONOxx.xxSTACKSxx.Domain.csproj
@@ -6,7 +6,6 @@
 
     <ItemGroup>
         <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Common\xxENSONOxx.xxSTACKSxx.Common.csproj" />
-        <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.csproj" />
     </ItemGroup>
 
     <ItemGroup>
