+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure/xxENSONOxx.xxSTACKSxx.Infrastructure.csproj
@@ -19,14 +19,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Application.CommandHandlers\xxENSONOxx.xxSTACKSxx.Application.CommandHandlers.csproj" />
     <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Application.Integration\xxENSONOxx.xxSTACKSxx.Application.Integration.csproj" />
-    <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Application.QueryHandlers\xxENSONOxx.xxSTACKSxx.Application.QueryHandlers.csproj" />
     <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.CQRS\xxENSONOxx.xxSTACKSxx.CQRS.csproj" />
     <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Domain\xxENSONOxx.xxSTACKSxx.Domain.csproj" />
     <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.csproj" />
     <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus.csproj" />
-    <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.csproj" />
 
   </ItemGroup>
