--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.Tests/FakeMenu.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.Tests;
-
-public class FakeMenu(Guid id)
-{
-	public FakeMenu() : this(Guid.NewGuid())
-    {
-    }
-
-    public Guid Id { get; } = id;
-}
