--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.DynamoDB/DynamoDbConfiguration.cs
@@ -1,7 +1,0 @@
-namespace xxENSONOxx.xxSTACKSxx.Shared.DynamoDB;
-
-public class DynamoDbConfiguration
-{
-	public string TableName { get; init; }
-	public string TablePrefix { get; init; }
-}
