+++ xxENSONOxx.xxSTACKSxx.sln
@@ -55,10 +55,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers.csproj", "{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.DynamoDB", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.csproj", "{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.Tests\xxENSONOxx.xxSTACKSxx.Shared.DynamoDB.Tests.csproj", "{8D556317-7C2B-445A-AE5E-92FF04969031}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions.csproj", "{7B827585-4F7A-473B-B994-681BE96B0C04}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Core", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Core\xxENSONOxx.xxSTACKSxx.Shared.Core.csproj", "{3F133394-CF28-4110-BA3D-EBF404208A81}"
@@ -433,30 +429,6 @@ Global
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|x64.Build.0 = Release|Any CPU
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|x86.ActiveCfg = Release|Any CPU
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|x86.Build.0 = Release|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Debug|x64.Build.0 = Debug|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Debug|x86.Build.0 = Debug|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Release|x64.ActiveCfg = Release|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Release|x64.Build.0 = Release|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Release|x86.ActiveCfg = Release|Any CPU
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5}.Release|x86.Build.0 = Release|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Debug|x64.Build.0 = Debug|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Debug|x86.Build.0 = Debug|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Release|x64.ActiveCfg = Release|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Release|x64.Build.0 = Release|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Release|x86.ActiveCfg = Release|Any CPU
-		{8D556317-7C2B-445A-AE5E-92FF04969031}.Release|x86.Build.0 = Release|Any CPU
 		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -787,8 +759,6 @@ Global
 		{7290B15D-B0DB-48FA-9A2A-1E8E15447CD2} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{51D0C76D-9F7F-4213-9BC1-0D7DE060E4AF} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
-		{3F45C828-B6C2-4D16-90DA-F8D034EC11D5} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
-		{8D556317-7C2B-445A-AE5E-92FF04969031} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{7B827585-4F7A-473B-B994-681BE96B0C04} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{3F133394-CF28-4110-BA3D-EBF404208A81} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{06DE9A76-034B-4975-8365-A7F88EF7954E} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
